requirements.txt
aiohttp==3.6.3
env==0.1.0
gcloud==0.18.3
googleapis-common-protos==1.53.0
imgurpython==1.1.7
praw==7.2.0
prawcore==2.0.0
pymongo==3.11.3
Pyrebase4==4.5.0
python-decouple==3.4
python-dotenv==0.17.1
python-environ==0.4.54
pytz==2019.1
requests==2.25.1
text-to-owo==1.1.6
TextToOwO==2.0.1
websocket-client==1.0.1
giphy_client
discord.py>=1.7.2
discord_components
urbandict